Bump to version 0.2.0
Contains many changes so that the script will become compatible with the
sh-checker GitHub Action.

Most notable of these changes are the quotation of most variables, the
switch from spaces to tabs in regards to indenting, and making use of
the branding variable so that it won't complain that it is not being
used. In regards to branding, it specifies what files to edit in order
to apply the branding.

Guidance on how to setup/upgrade Axelor was added, as it makes a useful
addition alongside the other changes made.

To become root is no longer required in order to download the source
code.