Move Axelor data to INI file
The Axelor data originally included with the ax-invconv script was based
off of an examination of Axelor Demo Data. Since the production version
doesn't contain the same categories and other data, and, as well, custom
data can be added to Axelor, all data has been moved to a seperate file.
This should allow for more flexibility in using the script.