Make parser more flexible
The script has been rewritten to allow it to continue in more
circumstances even if one file or section in a file wasn't parsed
properly.

The old identification system used for differentiating between files and
sections are now only used for error, warning, and info messages.
Internally, the script uses a tuple with both file name and worksheet
name for the greatest flexibility.