Add custom code support
Usually, codes are based off of an abbreviated form of Axelor Family
categories. As a result, get_code used to manipulate text from a given
cell.

From now on, when a column named "code" is in a input file and the
force-custom-code argument has been used, it will first try to use a code as
is.