get: Readd chown support
The original shell script chowned the folder resulting from the
extraction of the WAR file so that it would be owned by the tomcat user
and group. This feature has been readded in such a manner in order to
avoid compromising cross platform support.

Support for destination directories has also been added, since it is
a nice feature to introduce and permissions need to be taken care of.
Chowning files already requires root privileges, so the feature might as
well be introduced in the same commit, as (at least on POSIX-based
platforms) permissions are already taken care of.

Support for destination directories has also been added, since
it is a nice feature to add and permissions need to be taken care of, I
might as well add the feature in the same commit where I am adding support for
becoming root.

Some variables were renamed and more comments were added to make the
script easier to read.