invconv: Move arguments to data file
All of the arguments for fallback variables are now defined in the data
file. This will allow for greater flexibility and is in line with moving
the Axelor column names out of the main script.

The argument '--force-custom-code' has been removed, as it will be
superseded by new capabilities in the axm parser.