invconv: Turn axm into a programming language
A new axm package was added containing many new functions,
allowing for the addition of new operators and commands.

A new importer operator was introduced that parses user-defined strings
and includes these strings as part of the outputted CSV file. As a
result of this change, the invconv_logic functions are now based off of
reading the Axelor column names (as opposed to using the input column as
a reference).

The axm package optionally supports using functions from the msg_handler
module. The functions in that module have been extended to better
support axm.

All list appending now uses the append() method instead of the '+='
operator in order to make the code easier to read and to avoid mistakes.
The '+=' operator is also no longer used with strings as it is not
portable across Python implementations. As a result of these and other
changes, various potential problems were found and fixed.

The xlsx_headers dictionary no longer contains another dictionary as the
value of its keys in order to simplify the code and make it less
specific to the xlsx file format. This secondary dictionary contained
some 1-indexed position numbers, which made it easier to work with
openpyxl, but it is no longer neccessary due to the creation of the axm
module and the subsequent rewriting of the relevant code.