invconv: Get loguru to handle exceptions
This involved major refractoring, since a decorator is applied to a
function to enable reporting of exceptions for that function. Since the
entire script should report exceptions to the log file, a new main()
function was created. Since the main() function, as is, was too big,
handling of xlsx files was moved elsewhere and as a result, support for
other file types can easily be added.

Also as the result of the refractoring, the script can now be used as a
package in another script.