invconv: Allow new logic functions to be defined
The existing logic functions were moved to a new file. The changes
should result in more flexibility in output.

Also, the description and internal description related functions were
removed since due to the changes in v0.5.0 they are unnecessary. If
there isn't a function to manipulate output for a specific output
column, the input text will be used "as is". That is what the removed
functions did (simply returning the text they received).