Improve black.yml
Fixed black.yml so that it would use the latest black GitHub Action, but a specific version of the formatter itself. The src option was removed, as it wasn't needed, for black should be able to parse all files in the entire repository.

Also, the events that activate the workflow has been changed so that pull requests should no longer spawn this workflow more than once.