invconv: Fix axm regression
In version 0.5.0, a regression was introduced which caused default.axm
to no longer work. The problem was that the script expected both
out_input_col and column_output_dict to be non-empty. The former is used
to store the possible column mappings, and the latter is filled when the
importer operator is used (to assign an output column to a string and
not to a set of possible input columns). It is possible for either
dictionary to be empty while still having a proper axm file.