invconv: Fix column matching
If there was more than one column in the input that matched one of the
possible input columns for a specific output column in the axm file,
then instead of going with the first match it finds, it would go with
the last match it finds. The proper behaviour of going with the first
match has been restored.