invconv: Don't print ID when outputting to stream
The ID written by progress_bar.text() is not properly cleared when
replaced. As a result, pieces of the ID usually ends up in the CSV text
when outputting to stderr or stdout. This particular case has been
prevented.

Even when not outputting text to stderr/stdout, remnants of the previous
ID sometimes appear at the end of the current ID. That particular
cosmetic issue has not been fixed, since the ID still provides helpful
details to the user.